+++ Sources/LCP/License/License.swift
@@ -45,16 +45,31 @@ extension License: LCPLicense {
         documents.status
     }
 
+    public var isRestricted: Bool {
+        error != nil
+    }
+
+    public var error: StatusError? {
+        switch documents.context {
+        case .success:
+            return nil
+        case let .failure(error):
+            return error
+        }
+    }
+
     public var encryptionProfile: String? {
         license.encryption.profile
     }
 
     public func decipher(_ data: Data) throws -> Data? {
-        let context = try documents.getContext()
+        let context = try documents.context.get()
         return client.decrypt(data: data, using: context)
     }
 
     func charactersToCopyLeft() async -> Int? {
+        guard !isRestricted else { return 0 }
+
         do {
             return try await licenses.userRights(for: license.id).copy
         } catch {
@@ -64,6 +79,8 @@ extension License: LCPLicense {
     }
 
     func canCopy(text: String) async -> Bool {
+        guard !isRestricted else { return false }
+
         guard let charactersLeft = await charactersToCopyLeft() else {
             return true
         }
@@ -71,6 +88,8 @@ extension License: LCPLicense {
     }
 
     func copy(text: String) async -> Bool {
+        guard !isRestricted else { return false }
+
         do {
             var allowed = true
             try await licenses.updateUserRights(for: license.id) { rights in
@@ -94,6 +113,8 @@ extension License: LCPLicense {
     }
 
     func pagesToPrintLeft() async -> Int? {
+        guard !isRestricted else { return 0 }
+
         do {
             return try await licenses.userRights(for: license.id).print
         } catch {
@@ -103,6 +124,8 @@ extension License: LCPLicense {
     }
 
     func canPrint(pageCount: Int) async -> Bool {
+        guard !isRestricted else { return false }
+
         guard let pageLeft = await pagesToPrintLeft() else {
             return true
         }
@@ -110,6 +133,8 @@ extension License: LCPLicense {
     }
 
     func print(pageCount: Int) async -> Bool {
+        guard !isRestricted else { return false }
+
         do {
             var allowed = true
             try await licenses.updateUserRights(for: license.id) { rights in